chore: moving file structure of components
