--- src/components/Icon/index.js
@@ -1,2 +1,0 @@
-import Icon from "./Icon"
-export default Icon
