--- src/components/Layout/index.js
@@ -1,2 +1,0 @@
-import Layout from "./Layout"
-export default Layout
