--- src/components/Navbar/index.js
@@ -1,3 +1,0 @@
-import Navbar, { SimpleNavbar } from "./Navbar"
-export default Navbar
-export { SimpleNavbar }
