--- src/components/PrivateRoute/index.js
@@ -1,2 +1,0 @@
-import PrivateRoute from "./PrivateRoute"
-export default PrivateRoute
