--- src/components/Seek/index.js
@@ -1,2 +1,0 @@
-import Seek from "./Seek"
-export default Seek
