--- src/components/SignOut/index.js
@@ -1,2 +1,0 @@
-import SignOut from "./SignOut"
-export default SignOut
